Clock and bugfix
and some news: we tried to make printed circuit board with jakeph12
but it was unsuccessfull. Seems that DC-DC converter that I selected was not enough... Or I did circuit wrong.
Any way... I'll need to employ myself and make some money to continue developing hardware. Software will still be developed. Contributions are welcome.

Co-authored-by: jakeph12 <jakeph12@users.noreply.github.com>